capture: share a single session
