+++ cdc/owner_test.go
@@ -713,4 +713,9 @@ func (s *changefeedInfoSuite) TestMinimumTables(c *check.C) {
 	captures["c4"] = &model.CaptureInfo{}
 	c.Assert(cf.minimumTablesCapture(captures), check.Equals, "c4")
 }
+
+// TODO Test watchCapture
+func (s *ownerSuite) TestWatchCapture(c *check.C){
+	FIXME
+}
 */
