+++ cdc/server.go
@@ -24,6 +24,8 @@ import (
 	"go.uber.org/zap"
 )
 
+const ownerRunInterval = time.Millisecond * 500
+
 type options struct {
 	pdEndpoints string
 	statusHost  string
@@ -64,6 +66,7 @@ type ServerOption func(*options)
 type Server struct {
 	opts         options
 	capture      *Capture
+	owner        *Owner
 	statusServer *http.Server
 }
 
@@ -93,8 +96,44 @@ func NewServer(opt ...ServerOption) (*Server, error) {
 // Run runs the server.
 func (s *Server) Run(ctx context.Context) error {
 	s.startStatusHTTP()
-	ctx = util.PutCaptureIDInCtx(ctx, s.capture.info.ID)
-	return s.capture.Start(ctx)
+
+	// When a capture suicided, restart it
+	for {
+		if err := s.run(ctx); err != ErrSuicide {
+			return err
+		}
+	}
+}
+
+func (s *Server) run(ctx context.Context) (err error) {
+	ctx, cancel := context.WithCancel(util.PutCaptureIDInCtx(ctx, s.capture.info.ID))
+
+	// when a goroutine paniced, cancel would be called first, which
+	// cancels all the normal goroutines, and then the defered recover
+	// is called, which modifies the err value to ErrSuicide. The caller
+	// would restart this function when an error is ErrSuicide.
+	defer func() {
+		if r := recover(); r == ErrSuicide {
+			// assign the error value, which should be handled by
+			// the parent caller
+			err = ErrSuicide
+		}
+	}()
+	defer cancel()
+
+	go s.capture.Run(ctx)
+
+	// Campaign to be an owner, it blocks until it becomes
+	// the owner
+	if err := s.capture.Campaign(ctx); err != nil {
+		return err
+	}
+	owner, err := NewOwner(s.capture.session)
+	if err != nil {
+		return err
+	}
+	s.owner = owner
+	return owner.Run(ctx, ownerRunInterval)
 }
 
 // Close closes the server.
